+++ src/runner-nodejs.ts
@@ -1,13 +1,5 @@
 import { output } from './test';
 
-console.log('=== DEBUG LEVEL ===');
-output('debug', false);
+output('debug');
 
-console.log('=== INFO LEVEL ===');
-output('info', false);
-
-console.log('=== DEBUG LEVEL ===');
-output('debug', true);
-
-console.log('=== INFO LEVEL ===');
-output('info', true);
+output('info');